First draft of the node-analyzers plugin
It gets complete list of built-in analyzers, tokenizers, tokenFilters, chartFilters and normalizers from the node. On to of that it can list all available analysis plugins on that node and its registries.
Right now it has to use reflection to get set of keys from individual registries because these are private members of the AnalysisRegistry.

Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>